# docs/swagger.yaml
definitions:
...
  github_com_PritOriginal_problem-map-server_internal_models.KPI:
    properties:
      avg_close_hours:
        description: AvgCloseHours is the mean unconfirmed -> closed time; null when
          no mark was closed.
        type: number
      avg_confirm_hours:
        description: AvgConfirmHours is the mean unconfirmed -> confirmed time; null
          when no mark was confirmed.
        type: number
      by_organization:
        description: ByOrganization lists the assigned marks per organization.
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.OrganizationKPI'
        type: array
      by_status:
        additionalProperties:
          type: integer
        type: object
      median_confirm_hours:
        description: MedianConfirmHours is the median unconfirmed -> confirmed time;
          null when no mark was confirmed.
        type: number
      open_older_than_30d:
        description: |-
          OpenOlderThan30d counts marks neither closed nor refuted that were
          created more than 30 days ago.
        type: integer
      refuted_share:
        description: RefutedShare is refuted marks / total (0 when there are no marks).
        type: number
      sla_breach_share:
        description: |-
          SLABreachShare is the share of marks assigned to an organization whose
          SLA deadline has passed while they are still confirmed or in progress
          (0 when no mark is assigned).
        type: number
      total:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.OrganizationKPI:
    properties:
      name:
        type: string
      organization_id:
        type: integer
      overdue:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_PritOriginal_problem-map-server_internal_models.TimeseriesPoint:
    properties:
      closed:
        type: integer
      confirmed:
        type: integer
      created:
        type: integer
      period:
        type: string
      refuted:
        type: integer
    type: object
  github_com_PritOriginal_problem-map-server_internal_models.TopType:
    properties:
      count:
        type: integer
      mark_type_id:
        type: integer
      name:
        type: string
      share:
        type: number
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo:
    properties:
      message:
        type: string
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-any:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload: {}
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-github_com_PritOriginal_problem-map-server_internal_models_KPI:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.KPI'
      success:
        type: boolean
    type: object
...
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_analytics_GetTimeseriesResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_analytics.GetTimeseriesResponse'
      success:
        type: boolean
    type: object
  github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_analytics_GetTopTypesResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.ListMeta'
      payload:
        $ref: '#/definitions/internal_handler_analytics.GetTopTypesResponse'
      success:
        type: boolean
    type: object
...
  internal_handler_analytics.GetTimeseriesResponse:
    properties:
      timeseries:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.TimeseriesPoint'
        type: array
    type: object
  internal_handler_analytics.GetTopTypesResponse:
    properties:
      types:
        items:
          $ref: '#/definitions/github_com_PritOriginal_problem-map-server_internal_models.TopType'
        type: array
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the "Problem Map" project.
  title: Problem Map API
  version: "1.0"
paths:
...
  /analytics/kpi:
    get:
      description: Totals, per-status counts, confirmation/closing durations (hours,
        from the status history), refuted share and stale open marks. All filters
        are optional.
      parameters:
      - description: only marks inside this admin boundary
        in: query
        name: boundary_id
        type: integer
      - description: only marks of this type
        in: query
        name: mark_type_id
        type: integer
      - description: marks created at or after (RFC3339 or YYYY-MM-DD, taken as 00:00:00
          UTC)
        in: query
        name: from
        type: string
      - description: marks created at or before (RFC3339 or YYYY-MM-DD, taken as the
          end of that day in UTC)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-github_com_PritOriginal_problem-map-server_internal_models_KPI'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: KPI summary
      tags:
      - analytics
  /analytics/timeseries:
    get:
      description: 'Number of marks created and transitions to confirmed / closed
        / refuted per period; empty periods are returned with zeros. Defaults: step=day,
        to=now, from=to minus 30 days (12 weeks / 12 months for coarser steps).'
      parameters:
      - description: only marks inside this admin boundary
        in: query
        name: boundary_id
        type: integer
      - description: only marks of this type
        in: query
        name: mark_type_id
        type: integer
      - description: start of the range (RFC3339 or YYYY-MM-DD, taken as 00:00:00
          UTC)
        in: query
        name: from
        type: string
      - description: end of the range (RFC3339 or YYYY-MM-DD, taken as the end of
          that day in UTC)
        in: query
        name: to
        type: string
      - default: day
        description: bucket size
        enum:
        - day
        - week
        - month
        in: query
        name: step
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_analytics_GetTimeseriesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Marks time series
      tags:
      - analytics
  /analytics/top-types:
    get:
      description: Mark types ordered by the number of matching marks with their share
        of the total.
      parameters:
      - description: only marks inside this admin boundary
        in: query
        name: boundary_id
        type: integer
      - description: marks created at or after (RFC3339 or YYYY-MM-DD, taken as 00:00:00
          UTC)
        in: query
        name: from
        type: string
      - description: marks created at or before (RFC3339 or YYYY-MM-DD, taken as the
          end of that day in UTC)
        in: query
        name: to
        type: string
      - default: 10
        description: number of rows (1..100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-internal_handler_analytics_GetTopTypesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_PritOriginal_problem-map-server_pkg_responses.Response-any'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Top mark types
      tags:
      - analytics
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Open-data API key "pm_live_…" (POST /api-keys); also accepted as
      "Authorization: ApiKey {key}". Optional on public GET routes: with a key the
      per-key quota applies and X-RateLimit-* headers are returned; ignored when a
      valid Bearer JWT is sent as well (the JWT identity wins)'
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: 'JWT access token: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...